I found a bug, I have no idea why, I hate my life
I'm able to go back in time using github, but it's still extremely annoying. I believe it has something to do with me putting my main game loop inside of a class but the bug is also only affecting me when I have fullscreen, so at first I was thinking maybe it's something to do with memory but then why would a smaller window fix that.